Fix switch-bool warning
As of gcc version 5.1.1 a new warning has been added to detect the
use of a boolean in a switch statement (-Wswitch-bool).  Resolve the
warning by explicitly casting the value to an integer type.

  zfs-0.6.4/module/zfs/zvol.c: In function 'zvol_request':
  error: switch condition has boolean value [-Werror=switch-bool]

+++ module/zfs/zvol.c
@@ -764,7 +764,7 @@ zvol_request(struct request_queue *q)
 			continue;
 		}
 
-		switch (rq_data_dir(req)) {
+		switch ((int)rq_data_dir(req)) {
 		case READ:
 			zvol_dispatch(zvol_read, req);
 			break;
